feat: add network-error details tests for ECONNREFUSED and self-signed TLS errors
